+++ scripts/answers/script_adventure_6_1.py
@@ -1,4 +1,4 @@
-from vb import *
+from vx import *
 
 brk(3, 7)
 brk(3, 8)