+++ scripts/answers/script_adventure_6_3.py
@@ -1,5 +1,5 @@
 # Submit at 2026-08-15 18:08:46: 2113, 2133, 0
-from vb import *
+from vx import *
 
 for r in (2, 3, 4):
 	for c in (3, 4, 5, 6, 7):