+++ scripts/templates/script_adventure_6_1.py
@@ -1,3 +1,10 @@
-from vb import *
+from vx import *
 
 brk(3, 7)
+
+# Maybe there's one line
+# of code missing? Write
+# it down in another line:
+
+
+