+++ scripts/templates/script_adventure_6_4.py
@@ -1,3 +1,3 @@
-from vb import *
+from vx import *