+++ scripts/vb/__init__.py
@@ -1,2 +1,0 @@
-from .basic import *
-from .name import *